Minimal fixes to get example functioning again
