+++ examples/SuperResolution/voxel-ShapeNet/README.md
@@ -1,13 +1,13 @@
 # Example: voxel-superresolution
 
-These example scripts lets you train a voxel superresolution network on the ShapeNet Dataset using a simple encoder-decoder network. There are two training schemes, one using MSE loss and the other using negative log likelihood loss. 
+These example scripts let you train a voxel superresolution network on the ShapeNet Dataset using a simple encoder-decoder network. There are two training schemes, one using MSE loss and the other using negative log likelihood loss. 
 
 
 ### Training the network: MSE
 
 To train using MSE run
 ```
-python train_MSE.py
+python train.py --shapenet-root <ShapeNet dir> --cache <cache dir> --loss-type MSE
 ```
 
 
@@ -15,14 +15,14 @@ python train_MSE.py
 
 To evaluate a trained MSE model run 
 ```
-python eval_MSE.py
+python eval.py --shapenet-root <ShapeNet dir> --cache <cache dir> --loss-type MSE
 ```
 
 ### Training the network: NLLL
 
 To train using NLLL run
 ```
-python train_NLLL.py
+python train.py --shapenet-root <ShapeNet dir> --cache <cache dir> --loss-type MSE
 ```
 
 
@@ -30,5 +30,5 @@ python train_NLLL.py
 
 To evaluate a trained NLLL model run 
 ```
-python eval_NLLL.py
+python eval.py --shapenet-root <ShapeNet dir> --cache <cache dir> --loss-type MSE
 ```
