+++ examples/SuperResolution/voxel-ShapeNet/utils.py
@@ -1,10 +1,12 @@
-import torch 
+import torch
 
-import kaolin as kal
+import kaolin.transforms as tfs
 
-def up_sample(inp): 
-	NN_pred = []
-	for voxel in inp: 
-		NN_pred.append(kal.rep.voxel.scale_up(voxel, dim = 128))
-	NN_pred = torch.stack(NN_pred)
-	return NN_pred
+
+def up_sample(inp):
+    NN_pred = []
+    upsample_128 = tfs.UpsampleVoxelGrid(dim=128)
+    for voxel in inp:
+        NN_pred.append(upsample_128(voxel))
+    NN_pred = torch.stack(NN_pred)
+    return NN_pred
